+++ projects/spine/src/main/java/org/openmuc/jeebus/spine/impl/NodeManagementImpl.java
@@ -17,7 +17,9 @@ import org.openmuc.jeebus.spine.spi.*;
 import org.openmuc.jeebus.spine.spi.function.FeatureFunction;
 import org.openmuc.jeebus.spine.utils.NamedThreadFactory;
 import org.openmuc.jeebus.spine.xsd.v1.*;
+import org.openmuc.jeebus.spine.xsd.v1.NodeManagementBindingDeleteCallType.BindingDelete;
 import org.openmuc.jeebus.spine.xsd.v1.NodeManagementBindingRequestCallType.BindingRequest;
+import org.openmuc.jeebus.spine.xsd.v1.NodeManagementDetailedDiscoveryFeatureInformationType.Description;
 import org.openmuc.jeebus.spine.xsd.v1.NodeManagementSubscriptionDeleteCallType.SubscriptionDelete;
 import org.openmuc.jeebus.spine.xsd.v1.NodeManagementSubscriptionRequestCallType.SubscriptionRequest;
 import org.slf4j.Logger;
@@ -94,16 +96,17 @@ class NodeManagementImpl extends FeatureImpl implements NodeManagement {
         CmdType subscriptionDeleteCmd = new CmdType();
         NodeManagementSubscriptionDeleteCallType deleteCall
             = new NodeManagementSubscriptionDeleteCallType();
-        SubscriptionDelete subscription
-            = new SubscriptionDelete();
+        SubscriptionDelete subscription = new SubscriptionDelete();
         subscription.setClientAddress(clientAddress);
         subscription.setServerAddress(serverAddress);
         deleteCall.setSubscriptionDelete(subscription);
         subscriptionDeleteCmd.setNodeManagementSubscriptionDeleteCall(deleteCall);
+
         if (clientAddress.getDevice().equals(getDevice().getAddress().getDevice())) {
-            if (!serverAddress
-                .getDevice()
-                .equals(getDevice().getAddress().getDevice())) {
+            if (!Objects.equals(
+                serverAddress.getDevice(),
+                getDevice().getAddress().getDevice()
+            )) {
                 requestCall(
                     getNodeManagementAddress(serverAddress.getDevice()),
                     subscriptionDeleteCmd
@@ -118,9 +121,8 @@ class NodeManagementImpl extends FeatureImpl implements NodeManagement {
             );
         }
         try {
-            ((SubscriptionDataFunction) getFunction(
-                NODE_MANAGEMENT_SUBSCRIPTION_DATA)).removeSubscriptionEntry(
-                subscription);
+            ((SubscriptionDataFunction) getFunction(NODE_MANAGEMENT_SUBSCRIPTION_DATA))
+                .removeSubscriptionEntry(subscription);
         }
         catch (SpineException e) {
             LOGGER.error("There was an error trying to release a subscription:", e);
@@ -144,8 +146,8 @@ class NodeManagementImpl extends FeatureImpl implements NodeManagement {
         CmdType bindingDeleteCmd = new CmdType();
         NodeManagementBindingDeleteCallType bindingDeleteCall
             = new NodeManagementBindingDeleteCallType();
-        NodeManagementBindingDeleteCallType.BindingDelete bindingDelete
-            = new NodeManagementBindingDeleteCallType.BindingDelete();
+        BindingDelete bindingDelete
+            = new BindingDelete();
         bindingDelete.setClientAddress(clientAddress);
         bindingDelete.setServerAddress(serverAddress);
         bindingDeleteCall.setBindingDelete(bindingDelete);
@@ -365,14 +367,15 @@ class NodeManagementImpl extends FeatureImpl implements NodeManagement {
         BindingDataFunction bindingDataFunction
             = (BindingDataFunction) getFunction(
             NODE_MANAGEMENT_BINDING_DATA);
-        for (Map.Entry<String, List<NodeManagementBindingDeleteCallType.BindingDelete>> releasedBindings : bindingDataFunction
-            .deleteBindings(
-                address)
-            .entrySet()) {
+        for (Map.Entry<String, Set<BindingDelete>> releasedBindings :
+            bindingDataFunction
+                .deleteBindings(address)
+                .entrySet()
+        ) {
             CmdType bindingDeleteCmd = new CmdType();
             NodeManagementBindingDeleteCallType bindingDeleteCall
                 = new NodeManagementBindingDeleteCallType();
-            for (NodeManagementBindingDeleteCallType.BindingDelete bindingDelete : releasedBindings.getValue()) {
+            for (BindingDelete bindingDelete : releasedBindings.getValue()) {
                 bindingDeleteCall.setBindingDelete(bindingDelete);
                 bindingDeleteCmd.setNodeManagementBindingDeleteCall(bindingDeleteCall);
                 requestCall(
@@ -417,8 +420,8 @@ class NodeManagementImpl extends FeatureImpl implements NodeManagement {
     ) {
         NodeManagementDetailedDiscoveryFeatureInformationType featureInformation
             = new NodeManagementDetailedDiscoveryFeatureInformationType();
-        NodeManagementDetailedDiscoveryFeatureInformationType.Description description
-            = new NodeManagementDetailedDiscoveryFeatureInformationType.Description();
+        Description description
+            = new Description();
         description.setLastStateChange(change);
         description.setFeatureAddress(address);
         featureInformation.setDescription(description);
@@ -478,7 +481,8 @@ class NodeManagementImpl extends FeatureImpl implements NodeManagement {
     }
 
     public CmdType getBindingRequest(
-        FeatureAddressType clientAddress, FeatureAddressType serverAddress,
+        FeatureAddressType clientAddress,
+        FeatureAddressType serverAddress,
         FeatureTypeEnumType serverFeatureType
     ) {
         NodeManagementBindingRequestCallType bindingRequestCall
@@ -491,9 +495,6 @@ class NodeManagementImpl extends FeatureImpl implements NodeManagement {
 
         CmdType cmd = new CmdType();
         cmd.setNodeManagementBindingRequestCall(bindingRequestCall);
-
-        ((BindingDataFunction) getFunction(NODE_MANAGEMENT_BINDING_DATA)).addBinding(
-            bindingRequest);
 
         return cmd;
     }
@@ -536,7 +537,7 @@ class NodeManagementImpl extends FeatureImpl implements NodeManagement {
     private boolean containsDestinationListDataFunction(
         NodeManagementDetailedDiscoveryDataType detailedDiscoveryData
     ) {
-        NodeManagementDetailedDiscoveryFeatureInformationType.Description
+        Description
             nodeManagementDescription = getNodeManagementFeatureDescription(
             detailedDiscoveryData);
         if (nodeManagementDescription == null) {
@@ -592,7 +593,7 @@ class NodeManagementImpl extends FeatureImpl implements NodeManagement {
         }
     }
 
-    private NodeManagementDetailedDiscoveryFeatureInformationType.Description getNodeManagementFeatureDescription(
+    private Description getNodeManagementFeatureDescription(
         NodeManagementDetailedDiscoveryDataType detailedDiscoveryData
     ) {
         for (NodeManagementDetailedDiscoveryFeatureInformationType featureInfo
@@ -768,7 +769,8 @@ class NodeManagementImpl extends FeatureImpl implements NodeManagement {
     @Override
     public CompletableFuture<RequestResult> subscribeDestinationListData(
         String deviceId,
-        SpineSubscription subscription, boolean notifyFullRead
+        SpineSubscription subscription,
+        boolean notifyFullRead
     ) {
         CompletableFuture<RequestResult> future = requestSubscription(
             getNodeManagementAddress(deviceId),
@@ -812,10 +814,8 @@ class NodeManagementImpl extends FeatureImpl implements NodeManagement {
 
     @Override
     public void removeAddressMapping(String communicationAddress) {
-        DestinationListDataFunction function
-            = (DestinationListDataFunction) getFunction(
-            NODE_MANAGEMENT_DESTINATION_LIST_DATA);
-        function.deleteAddressMapping(communicationAddress);
+        getFunction(DestinationListDataFunction.class)
+            .ifPresent(function -> function.deleteAddressMapping(communicationAddress));
     }
 
     @Override
@@ -823,7 +823,16 @@ class NodeManagementImpl extends FeatureImpl implements NodeManagement {
         DisconnectReason disconnectReason,
         ShipConnectionInterface shipConnectionInterface
     ) {
+        notifyDisconnect(shipConnectionInterface.getRemoteId());
+    }
+
+    @Override
+    public void notifyDisconnect(String deviceAddress) {
         // TODO: stop use case execution on disconnected remote devices
+
+        getFunction(SubscriptionDataFunction.class)
+            .orElseThrow()
+            .removeSubscription(deviceAddress);
     }
 
     @Override