+++ projects/spine/src/test/java/org/openmuc/jeebus/spine/impl/SubscriptionTest.java
@@ -8,9 +8,12 @@
  * SPDX-License-Identifier: EPL-2.0
  ********************************************************************************/
 
-package org.openmuc.jeebus.spine;
+package org.openmuc.jeebus.spine.impl;
 
+import org.hamcrest.MatcherAssert;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.parallel.Execution;
 import org.openmuc.jeebus.spine.api.Device;
@@ -26,12 +29,14 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.stream.Collectors;
 
-import static java.util.concurrent.TimeUnit.SECONDS;
-import static org.awaitility.Awaitility.await;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
 import static org.junit.jupiter.api.parallel.ExecutionMode.SAME_THREAD;
 import static org.openmuc.jeebus.spine.TestUtilities.*;
+import static org.openmuc.jeebus.spine.impl.SubscriptionWrapper.State.SUCCESSFUL;
+import static org.openmuc.jeebus.spine.xsd.v1.FeatureTypeEnumType.GENERIC;
+import static org.openmuc.jeebus.spine.xsd.v1.RoleType.CLIENT;
+import static org.openmuc.jeebus.spine.xsd.v1.RoleType.SERVER;
 
 @Execution(SAME_THREAD)
 public class SubscriptionTest {
@@ -47,7 +52,7 @@ public class SubscriptionTest {
                 REMOTE_COMM,
                 REMOTE_DEVICE_ADDRESS
             ),
-            RoleType.SERVER
+            SERVER
         ).withDiscoverDevices(false).build();
 
         client = addFeature(
@@ -55,7 +60,7 @@ public class SubscriptionTest {
                 LOCAL_COMM,
                 LOCAL_DEVICE_ADDRESS
             ),
-            RoleType.CLIENT
+            CLIENT
         ).withDiscoverDevices(false).build();
     }
 
@@ -65,7 +70,7 @@ public class SubscriptionTest {
         client.getFeature(CLIENT_FEATURE_ADDRESS)
             .requestSubscription(
                 SERVER_FEATURE_ADDRESS,
-                FeatureTypeEnumType.GENERIC,
+                GENERIC,
                 SUBSCRIPTION
             );
 
@@ -80,7 +85,7 @@ public class SubscriptionTest {
             .getFeature(CLIENT_FEATURE_ADDRESS)
             .requestSubscription(
                 SERVER_FEATURE_ADDRESS,
-                FeatureTypeEnumType.GENERIC,
+                GENERIC,
                 SUBSCRIPTION
             );
 
@@ -88,7 +93,7 @@ public class SubscriptionTest {
             .getFeature(CLIENT_FEATURE_ADDRESS)
             .requestSubscription(
                 SERVER_FEATURE_ADDRESS,
-                FeatureTypeEnumType.GENERIC,
+                GENERIC,
                 new AssertingSubscription()
             );
 
@@ -137,6 +142,7 @@ public class SubscriptionTest {
         Device deviceUnderTest,
         boolean compare
     ) throws InterruptedException, ExecutionException {
+
         CompletableFuture<RequestResult> subscriptionReadRequest = requestingDevice
             .getNodeManagement()
             .requestSubscriptionData(deviceUnderTest.getAddress().getDevice());
@@ -146,9 +152,10 @@ public class SubscriptionTest {
             .getCmd()
             .getNodeManagementSubscriptionData();
 
-        await()
-            .atMost(1, SECONDS)
-            .until(() -> containsSubscription(subscriptionData) == compare);
+        MatcherAssert.assertThat(
+            containsSubscription(subscriptionData),
+            is(compare)
+        );
     }
 
     private boolean containsSubscription(NodeManagementSubscriptionDataType subscriptionData) {
@@ -184,7 +191,7 @@ public class SubscriptionTest {
         client.getFeature(CLIENT_FEATURE_ADDRESS)
             .requestSubscription(
                 SERVER_FEATURE_ADDRESS,
-                FeatureTypeEnumType.GENERIC,
+                GENERIC,
                 SUBSCRIPTION
             );
 
@@ -200,11 +207,10 @@ public class SubscriptionTest {
         client.getFeature(CLIENT_FEATURE_ADDRESS)
             .requestSubscription(
                 SERVER_FEATURE_ADDRESS,
-                FeatureTypeEnumType.GENERIC,
+                GENERIC,
                 SUBSCRIPTION
             );
 
-        //FIXME sometimes fails with ConcurrentModificationException
         client
             .getFeature(CLIENT_FEATURE_ADDRESS)
             .releaseSubscription(SERVER_FEATURE_ADDRESS);
@@ -218,7 +224,7 @@ public class SubscriptionTest {
         client.getFeature(CLIENT_FEATURE_ADDRESS)
             .requestSubscription(
                 SERVER_FEATURE_ADDRESS,
-                FeatureTypeEnumType.GENERIC,
+                GENERIC,
                 SUBSCRIPTION
             );
         server
@@ -238,11 +244,10 @@ public class SubscriptionTest {
         client.getFeature(CLIENT_FEATURE_ADDRESS)
             .requestSubscription(
                 SERVER_FEATURE_ADDRESS,
-                FeatureTypeEnumType.GENERIC,
+                GENERIC,
                 SUBSCRIPTION
             );
 
-        //FIXME: this fails (rarely) with a ConcurrentModificationException
         server.getEntity(1).deleteFeature(0);
         assertDeviceRegisteredSubscription(server, client, false);
         assertDeviceRegisteredSubscription(client, server, false);
@@ -254,13 +259,106 @@ public class SubscriptionTest {
         client.getFeature(CLIENT_FEATURE_ADDRESS)
             .requestSubscription(
                 SERVER_FEATURE_ADDRESS,
-                FeatureTypeEnumType.GENERIC,
+                GENERIC,
                 SUBSCRIPTION
             );
 
         client.getEntity(1).deleteFeature(0);
         assertDeviceRegisteredSubscription(server, client, false);
         assertDeviceRegisteredSubscription(client, server, false);
+    }
+
+    @Test
+    public void testSubscriptionCreationByWrapper() throws SpineException {
+        client.getFeature(CLIENT_FEATURE_ADDRESS)
+            .requestSubscription(
+                SERVER_FEATURE_ADDRESS,
+                GENERIC,
+                SUBSCRIPTION
+            ).join();
+
+        SubscriptionWrapper wrapper = ((FeatureImpl) client
+            .getFeature(CLIENT_FEATURE_ADDRESS))
+            .getSubscription(SERVER_FEATURE_ADDRESS);
+
+        assertThat(
+            wrapper.getState(),
+            is(SUCCESSFUL)
+        );
+        assertThat(
+            wrapper.getSubscriptions(),
+            hasSize(1)
+        );
+        assertThat(
+            wrapper.getSubscriptions(),
+            hasItem(SUBSCRIPTION)
+        );
+
+    }
+
+    @Test
+    public void testSubscriptionDeletionOnDisconnect() throws SpineException {
+        client.getFeature(CLIENT_FEATURE_ADDRESS)
+            .requestSubscription(
+                SERVER_FEATURE_ADDRESS,
+                GENERIC,
+                SUBSCRIPTION
+            ).join();
+
+        assertThat(
+            ((FeatureImpl) client.getFeature(CLIENT_FEATURE_ADDRESS))
+                .getSubscription(SERVER_FEATURE_ADDRESS),
+            is(notNullValue())
+        );
+
+        server.close();
+        assertThat(
+            client
+                .getDevice()
+                .getNodeManagement()
+                .getFunction(FunctionEnumType.NODE_MANAGEMENT_SUBSCRIPTION_DATA)
+                .read(null, SERVER_FEATURE_ADDRESS)
+                .getNodeManagementSubscriptionData()
+                .getSubscriptionEntry(),
+            hasSize(0)
+        );
+
+        assertThat(
+            ((FeatureImpl) client.getFeature(CLIENT_FEATURE_ADDRESS))
+                .getSubscription(SERVER_FEATURE_ADDRESS),
+            is(nullValue())
+        );
+    }
+
+    @Test
+    public void testDoubleSubscriptionByWrapper() throws SpineException {
+
+        CompletableFuture.allOf(
+            client.getFeature(CLIENT_FEATURE_ADDRESS)
+                .requestSubscription(
+                    SERVER_FEATURE_ADDRESS,
+                    GENERIC,
+                    SUBSCRIPTION
+                ),
+            client.getFeature(CLIENT_FEATURE_ADDRESS)
+                .requestSubscription(
+                    SERVER_FEATURE_ADDRESS,
+                    GENERIC,
+                    notification -> {}
+                )
+        ).join();
+
+        assertThat(
+            ((FeatureImpl) client.getFeature(CLIENT_FEATURE_ADDRESS))
+                .getSubscription(SERVER_FEATURE_ADDRESS),
+            is(notNullValue())
+        );
+
+        assertThat(
+            ((FeatureImpl) client.getFeature(CLIENT_FEATURE_ADDRESS))
+                .getSubscription(SERVER_FEATURE_ADDRESS).getSubscriptions(),
+            hasSize(2)
+        );
     }
 
 }